move filestore
